requirements.txt
opencv-python>=4.5.0
mediapipe>=0.10.0
numpy>=1.20.0
pyserial>=3.5